.backup/scripts/requirements.txt
actionlib==1.12.1
angles==1.9.12
bondpy==1.8.5
catboost==1.2.8
catkin==0.7.29
contourpy==1.3.1
cv_bridge==1.13.1
cycler==0.12.1
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
dynamic_reconfigure==1.6.5
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
graphviz==0.20.3
imbalanced-learn==0.13.0
interactive-markers==1.11.5
Jinja2==3.1.6
joblib==1.4.2
joint_state_publisher==1.12.15
kdl-parser-py==1.13.3
kiwisolver==1.4.8
laser_geometry==1.6.7
lightgbm==4.6.0
MarkupSafe==3.0.2
matplotlib==3.10.1
message_filters==1.14.13
mpmath==1.3.0
narwhals==1.35.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
psutil==7.0.0
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python_qt_binding==0.4.4
pytz==2025.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource_retriever==1.12.7
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospy==1.14.13
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.3
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rviz==1.13.30
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sensor-msgs==1.12.8
setuptools==79.0.0
six==1.17.0
sklearn-compat==0.1.3
smach==2.0.3
smach_ros==2.0.3
smclib==1.8.5
sympy==1.13.1
tf==1.12.1
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
threadpoolctl==3.6.0
topic_tools==1.14.13
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
typing_extensions==4.13.1
tzdata==2025.2
urdfdom-py==0.4.6
xacro==1.13.20
